Update config schema
