--- services/newsfeed/config.global.js
@@ -1,23 +1,0 @@
-/*
- * LiskHQ/lisk-service
- * Copyright © 2021 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-const config = {};
-config.log = {};
-config.electrum = {};
-config.interfaces = {};
-config.interfaces.request = {};
-config.interfaces.subscribe = {};
-
-module.exports = config;
